Adjustments made to script to allow user to run script themselves:
sp needs to load the species params object with n,p,d, alpha etc. included for newAllometric params to work, current data loaded by sp does not allow for this.

Added adjustments to the ecopath_catch data. Needs the following adjustments made to the data to be compatible with the AddEcopathCatchTotal() function.